# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dapp",
        "description": "Dapp application rest api",
        "contact": {
            "email": "babichev.oleg.n@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/glossarycard": {
            "post": {
                "tags": [
                    "Glossary card"
                ],
                "summary": "Create glossary card",
                "description": "Create glossary card",
                "operationId": "glossarycardPost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "word": {
                                        "type": "integer"
                                    },
                                    "translation": {
                                        "type": "integer"
                                    },
                                    "glossary": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "word": 100,
                                    "translation": 100,
                                    "glossary": 4
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/glossarycard/{glossarycardId}": {
            "delete": {
                "tags": [
                    "Glossary card"
                ],
                "summary": "Delete glossary card",
                "description": "Delete glossary card",
                "operationId": "delGlossaryCard",
                "parameters": [
                    {
                        "name": "glossarycardId",
                        "in": "path",
                        "description": "ID of glossary card",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/glossary": {
            "get": {
                "tags": [
                    "Glossary"
                ],
                "summary": "Get glossaries list",
                "description": "Returns glossaries list",
                "operationId": "getGlossaries",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Glossary"
                ],
                "summary": "Create glossary",
                "description": "Create glossary",
                "operationId": "glossaryPost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "unknown"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/glossary/{glossaryId}": {
            "get": {
                "tags": [
                    "Glossary"
                ],
                "summary": "Get glossary",
                "description": "Returns glossary",
                "operationId": "getGlossary",
                "parameters": [
                    {
                        "name": "glossaryId",
                        "in": "path",
                        "description": "ID of glossary",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Glossary"
                ],
                "summary": "Create glossary",
                "description": "Create glossary",
                "operationId": "glossaryPost",
                "parameters": [
                    {
                        "name": "glossaryId",
                        "in": "path",
                        "description": "ID of glossary",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "unknown"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Glossary"
                ],
                "summary": "Delete glossary",
                "description": "Delete glossary",
                "operationId": "delGlossary",
                "parameters": [
                    {
                        "name": "glossaryId",
                        "in": "path",
                        "description": "ID of glossary",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/translation/search": {
            "post": {
                "tags": [
                    "Translation"
                ],
                "summary": "Search translation by wordId and value",
                "description": "Search translation by wordId and value",
                "operationId": "searchTranslation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "val": {
                                        "type": "string"
                                    },
                                    "word_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "word_id": 403102,
                                    "val": ""
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}